# swagger-br.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Store Manager",
    "description": "Esta API é um sistema de gerenciamento de produtos e vendas",
    "contact": {
      "email": "raphael.almeida.martins@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "API de teste"
    }
  ],
  "paths": {
    "/products": {
      "get": {
        "summary": "Listagem de produto",
        "description": "Esta rota é responsável por listar todos os produtos cadastrados",
        "tags": ["Products"],
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                },
                "examples": {
                  "products": {
                    "value": [
                      {
                        "id": 1,
                        "name": "Martelo de Thor"
                      },
                      {
                        "id": 2,
                        "name": "Traje de encolhimento"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cadastrar um produto",
        "description": "Esta rota é responsável por cadastrar um novo produto",
        "tags": ["Products"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterProduct"
              },
              "examples": {
                "product": {
                  "value": {
                    "name": "Esmeralda do Chaos"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                },
                "examples": {
                  "product": {
                    "value": {
                      "id": 4,
                      "name": "Esmeralda do Chaos"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "\"name\" is required"
          },
          "422": {
            "description": "\"name\" is invalid"
          }
        }
      }
    },
...
    "/sales": {
      "get": {
        "summary": "Listagem de vendas",
        "description": "Esta rota é responsável por listar todas as vendas cadastradas",
        "tags": ["Sales"],
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSales"
                },
                "examples": {
                  "sales": {
                    "value": [
                      {
                        "saleId": 1,
                        "date": "2021-09-09T04:54:29.000Z",
                        "productId": 1,
                        "quantity": 2
                      },
                      {
                        "saleId": 1,
                        "date": "2021-09-09T04:54:54.000Z",
                        "productId": 2,
                        "quantity": 2
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cadastra venda",
        "description": "Esta rota é responsável por cadastrar uma nova venda",
        "tags": ["Sales"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterSale"
              },
              "examples": {
                "category": {
                  "value": [
                    {
                      "productId": 1,
                      "quantity": 1
                    },
                    {
                      "productId": 2,
                      "quantity": 5
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sale"
                },
                "examples": {
                  "sale": {
                    "value": {
                      "id": 3,
                      "itemsSold": [
                        {
                          "productId": 1,
                          "quantity": 1
                        },
                        {
                          "productId": 2,
                          "quantity": 5
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Some fields are invalid or missing"
          },
          "422": {
            "description": "\"quantity\" must be greater than or equal to 1"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "RegisterProduct": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "ListSales": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "saleId": {
              "type": "number"
            },
            "date": {
              "type": "string"
            },
            "productId": {
              "type": "number"
            },
            "quantity": {
              "type": "number"
            }
          }
        }
      },
...
      "RegisterSale": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "productId": {
              "type": "number"
            },
            "quantity": {
              "type": "number"
            }
          }
        }
      },
      "Sale": {
        "properties": {
          "id": {
            "type": "number"
          },
          "itemsSold": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": {
                  "type": "number"
                },
                "quantity": {
                  "type": "number"
                }
              }
            }
          }
        }
      },
...
    }
  }
}